+++ src/main/java/thedarkcolour/exdeorum/client/ter/SieveRenderer.java
@@ -18,9 +18,6 @@
 
 package thedarkcolour.exdeorum.client.ter;
 
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
@@ -34,15 +31,11 @@ import net.minecraftforge.registries.ForgeRegistries;
 import thedarkcolour.exdeorum.blockentity.SieveBlockEntity;
 import thedarkcolour.exdeorum.client.RenderUtil;
 
+import java.util.IdentityHashMap;
+import java.util.Map;
+
 public class SieveRenderer implements BlockEntityRenderer<SieveBlockEntity> {
-    public static final LoadingCache<Item, TextureAtlasSprite> MESH_TEXTURES = CacheBuilder.newBuilder().build(new CacheLoader<>() {
-        @Override
-        public TextureAtlasSprite load(Item key) {
-            ResourceLocation registryName = ForgeRegistries.ITEMS.getKey(key);
-            var textureLoc = registryName.withPrefix("item/mesh/");
-            return RenderUtil.blockAtlas.getSprite(textureLoc);
-        }
-    });
+    public static final Map<Item, TextureAtlasSprite> MESH_TEXTURES = new IdentityHashMap<>();
 
     public SieveRenderer(BlockEntityRendererProvider.Context ctx) {}
 
@@ -53,17 +46,26 @@ public class SieveRenderer implements BlockEntityRenderer<SieveBlockEntity> {
         if (!contents.isEmpty() && contents.getItem() instanceof BlockItem blockItem) {
             var block = blockItem.getBlock();
             var percentage = (float) sieve.getProgress() / 100.0f;
-            var builder = buffers.getBuffer(RenderUtil.TOP_RENDER_TYPES.getUnchecked(block));
-            var sprite = RenderUtil.TOP_TEXTURES.getUnchecked(block);
-
-            RenderUtil.renderFlatSpriteLerp(builder, stack, percentage, 0xff, 0xff, 0xff, sprite, light, 1.0f, 13f, 15f);
+            var face = RenderUtil.getTopFace(block);
+            face.renderFlatSpriteLerp(buffers, stack, percentage, 0xff, 0xff, 0xff, light, 1.0f, 13f, 15f);
         }
 
         var mesh = sieve.getMesh();
 
         if (!mesh.isEmpty()) {
             var builder = buffers.getBuffer(RenderType.cutoutMipped());
-            var sprite = MESH_TEXTURES.getUnchecked(mesh.getItem());
+            var meshItem = mesh.getItem();
+
+            TextureAtlasSprite sprite;
+            if (MESH_TEXTURES.containsKey(meshItem)) {
+                sprite = MESH_TEXTURES.get(meshItem);
+            } else {
+                ResourceLocation registryName = ForgeRegistries.ITEMS.getKey(meshItem);
+                ResourceLocation textureLoc = registryName.withPrefix("item/mesh/");
+                sprite = RenderUtil.blockAtlas.getSprite(textureLoc);
+                MESH_TEXTURES.put(meshItem, sprite);
+            }
+
             RenderUtil.renderFlatSprite(builder, stack, 0.75f, 0xff, 0xff, 0xff, sprite, light, 1f);
 
             if (mesh.hasFoil()) {